# docs/api/pic-sure-auth-services.openapi.json
{
  "components" : {
    "schemas" : {
      "AccessRule" : {
        "properties" : {
          "checkMapKeyOnly" : {
            "type" : "boolean"
          },
          "checkMapNode" : {
            "type" : "boolean"
          },
          "description" : {
            "type" : "string"
          },
          "evaluateOnlyByGates" : {
            "type" : "boolean"
          },
          "gateAnyRelation" : {
            "type" : "boolean"
          },
          "gates" : {
            "items" : {
              "$ref" : "#/components/schemas/AccessRule"
            },
            "type" : "array",
            "uniqueItems" : true
          },
          "mergedName" : {
            "type" : "string"
          },
          "mergedValues" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array",
            "uniqueItems" : true
          },
          "name" : {
            "type" : "string"
          },
          "rule" : {
            "type" : "string"
          },
          "subAccessRule" : {
            "items" : {
              "$ref" : "#/components/schemas/AccessRule"
            },
            "type" : "array",
            "uniqueItems" : true
          },
          "type" : {
            "format" : "int32",
            "type" : "integer"
          },
          "uuid" : {
            "format" : "uuid",
            "type" : "string"
          },
          "value" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ApplicationForDisplay" : {
        "properties" : {
          "description" : {
            "type" : "string"
          },
          "enable" : {
            "type" : "boolean"
          },
          "name" : {
            "type" : "string"
          },
          "url" : {
            "type" : "string"
          },
          "uuid" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "Connection" : {
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "label" : {
            "type" : "string"
          },
          "requiredFields" : {
            "type" : "string"
          },
          "subPrefix" : {
            "type" : "string"
          },
          "uuid" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "Privilege" : {
        "properties" : {
          "accessRules" : {
            "items" : {
              "$ref" : "#/components/schemas/AccessRule"
            },
            "type" : "array",
            "uniqueItems" : true
          },
          "application" : {
            "$ref" : "#/components/schemas/ApplicationForDisplay"
          },
          "description" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "uuid" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "Role" : {
        "properties" : {
          "description" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "privileges" : {
            "items" : {
              "$ref" : "#/components/schemas/Privilege"
            },
            "type" : "array",
            "uniqueItems" : true
          },
          "uuid" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "User" : {
        "properties" : {
          "acceptedTOS" : {
            "format" : "date-time",
            "type" : "string"
          },
          "active" : {
            "type" : "boolean"
          },
          "auth0metadata" : {
            "type" : "string"
          },
          "connection" : {
            "$ref" : "#/components/schemas/Connection"
          },
          "email" : {
            "type" : "string"
          },
          "generalMetadata" : {
            "type" : "string"
          },
          "matched" : {
            "type" : "boolean"
          },
          "passport" : {
            "type" : "string"
          },
          "roles" : {
            "items" : {
              "$ref" : "#/components/schemas/Role"
            },
            "type" : "array",
            "uniqueItems" : true
          },
          "subject" : {
            "type" : "string"
          },
          "token" : {
            "type" : "string"
          },
          "uuid" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "UserForDisplay" : {
        "properties" : {
          "acceptedTOS" : {
            "type" : "boolean"
          },
          "email" : {
            "type" : "string"
          },
          "privileges" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array",
            "uniqueItems" : true
          },
          "token" : {
            "type" : "string"
          },
          "uuid" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/user" : {
      "get" : {
        "description" : "GET a list of existing users, requires ADMIN or SUPER_ADMIN roles",
        "operationId" : "getUserAll",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/User"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [
          "User Management"
        ]
      },
      "post" : {
        "description" : "POST a list of users, requires ADMIN role",
        "operationId" : "addUser",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "description" : "A list of user in JSON format",
                "items" : {
                  "$ref" : "#/components/schemas/User"
                },
                "type" : "array"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/User"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [
          "User Management"
        ]
      },
      "put" : {
        "description" : "Update a list of users, will only update the fields listed, requires ADMIN role",
        "operationId" : "updateUser",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "items" : {
                  "$ref" : "#/components/schemas/User"
                },
                "type" : "array"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/User"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [
          "User Management"
        ]
      }
    },
    "/user/me" : {
      "get" : {
        "description" : "Retrieve information of current user",
        "operationId" : "getCurrentUser",
        "parameters" : [
          {
            "in" : "header",
            "name" : "Authorization",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "description" : "Attribute that represents if a long term token will attach to the response",
            "in" : "query",
            "name" : "hasToken",
            "required" : false,
            "schema" : {
              "type" : "boolean"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserForDisplay"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [
          "User Management"
        ]
      }
    },
...
    "/user/{userId}" : {
      "get" : {
        "description" : "GET information of one user with the UUID, requires ADMIN or SUPER_ADMIN roles",
        "operationId" : "getUserById",
        "parameters" : [
          {
            "description" : "The UUID of the user to fetch information about",
            "in" : "path",
            "name" : "userId",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [
          "User Management"
        ]
      }
    }
  },
  "servers" : [
    {
      "description" : "Generated server url",
      "url" : "http://localhost"
    }
  ]
}